fix store controller

+++ src/Controllers/StoreController.cs
@@ -20,6 +20,7 @@ public class StoreController(AppDbContext context, IHttpContextAccessor accessor
     [HttpPost("order"), Authorize]
     [SwaggerOperation("place an order. you must be logged in to do so")]
     [SwaggerResponse(StatusCodes.Status200OK, "ok")]
+    [SwaggerResponse(StatusCodes.Status400BadRequest, "pet with this id was already ordered", typeof(void))]
     [SwaggerResponse(StatusCodes.Status404NotFound, "pet with this id does not exist", typeof(void))]
     [SwaggerResponse(StatusCodes.Status401Unauthorized, "you must be logged in to place an order", typeof(void))]
     public async Task<IActionResult> Order([Required] uint petId)
@@ -30,21 +31,26 @@ public class StoreController(AppDbContext context, IHttpContextAccessor accessor
         if (pet == null || user == null)
             return NotFound();
 
+        if (pet.Status != PetStatus.Available)
+            return BadRequest();
+
         var order = new Order
         {
             PetId = petId,
             UserId = userId,
             Status = OrderStatus.Placed,
-            PlacedDate = DateTime.Now
+            PlacedDate = DateTime.Now.ToUniversalTime()
         };
+        pet.Status = PetStatus.Pending;
         await context.Orders.AddAsync(order);
         await context.SaveChangesAsync();
         return Ok();
     }
 
-    [HttpDelete("order/{petId}"), Authorize]
+    [HttpDelete("order/{orderId}"), Authorize]
     [SwaggerOperation("remove an order. You must be the user who placed the order")]
     [SwaggerResponse(StatusCodes.Status200OK, "ok")]
+    [SwaggerResponse(StatusCodes.Status400BadRequest, "a completed order cannot be canceled", typeof(void))]
     [SwaggerResponse(StatusCodes.Status404NotFound, "order with this id does not exist", typeof(void))]
     [SwaggerResponse(StatusCodes.Status401Unauthorized, "you must be logged in to delete an order", typeof(void))]
     [SwaggerResponse(StatusCodes.Status403Forbidden, "this is not your order", typeof(void))]
@@ -57,21 +63,28 @@ public class StoreController(AppDbContext context, IHttpContextAccessor accessor
             return NotFound();
         if (userId != order.UserId)
             return Forbid();
+        if (order.Status == OrderStatus.Completed)
+            return BadRequest();
 
+        var pet = await context.Pets.FindAsync(order.PetId);
+        if (pet == null)
+            return BadRequest();
+
+        pet.Status = PetStatus.Available;
         context.Orders.Remove(order);
         await context.SaveChangesAsync();
         return Ok();
     }
 
-    [HttpGet("order/{petId}"), Authorize]
+    [HttpGet("order/{orderId}"), Authorize]
     [SwaggerOperation("view an order. You must either be the user who placed the order or an admin")]
     [SwaggerResponse(StatusCodes.Status200OK, "ok", typeof(Order))]
     [SwaggerResponse(StatusCodes.Status404NotFound, "order with this id does not exist", typeof(void))]
     [SwaggerResponse(StatusCodes.Status401Unauthorized, "you must be logged in to view an order", typeof(void))]
     [SwaggerResponse(StatusCodes.Status403Forbidden, "this is not your order", typeof(void))]
-    public async Task<IActionResult> Get(uint petId)
+    public async Task<IActionResult> Get(uint orderId)
     {
-        var order = await context.Orders.FindAsync(petId);
+        var order = await context.Orders.FindAsync(orderId);
         if (order == null)
             return NotFound();
 
@@ -81,15 +94,15 @@ public class StoreController(AppDbContext context, IHttpContextAccessor accessor
         return Ok(order);
     }
 
-    [HttpPut("order/advance/{petId}"), Authorize(Roles = "Admin")]
+    [HttpPut("order/advance/{orderId}"), Authorize(Roles = "Admin")]
     [SwaggerOperation("advance an order")]
     [SwaggerResponse(StatusCodes.Status200OK, "ok")]
     [SwaggerResponse(StatusCodes.Status404NotFound, "order with this id does not exist", typeof(void))]
     [SwaggerResponse(StatusCodes.Status401Unauthorized, "you must be logged in to advance an order", typeof(void))]
     [SwaggerResponse(StatusCodes.Status403Forbidden, "you must be an admin to advance an order", typeof(void))]
-    public async Task<IActionResult> Advance(uint petId)
+    public async Task<IActionResult> Advance(uint orderId)
     {
-        var order = await context.Orders.FindAsync(petId);
+        var order = await context.Orders.FindAsync(orderId);
         if (order == null)
             return NotFound();
 
@@ -98,9 +111,14 @@ public class StoreController(AppDbContext context, IHttpContextAccessor accessor
         else if (order.Status == OrderStatus.Approved)
         {
             order.Status = OrderStatus.Completed;
-            order.CompletedDate = DateTime.Now;
+            order.CompletedDate = DateTime.Now.ToUniversalTime();
+            var pet = await context.Pets.FindAsync(order.PetId);
+            if (pet == null)
+            {
+                return BadRequest();
+            }
+            pet.Status = PetStatus.Sold;
         }
-
         await context.SaveChangesAsync();
         return Ok();
     }
@@ -113,5 +131,15 @@ public class StoreController(AppDbContext context, IHttpContextAccessor accessor
     public async Task<IActionResult> List([Required, FromQuery] HashSet<OrderStatus> status)
     {
         return Ok(await context.Orders.Where(u => status.Contains(u.Status)).ToListAsync());
+    }
+
+    [HttpGet("my_orders"), Authorize()]
+    [SwaggerOperation("list all orders of a user")]
+    [SwaggerResponse(StatusCodes.Status200OK, "ok")]
+    [SwaggerResponse(StatusCodes.Status401Unauthorized, "you must be logged in to advance an order", typeof(void))]
+    public async Task<IActionResult> ListMy([Required, FromQuery] HashSet<OrderStatus> status)
+    {
+        var id = claims.Id();
+        return Ok(await context.Orders.Where(u => status.Contains(u.Status) && u.UserId == id).ToListAsync());
     }
 }
